backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
google-genai
python-multipart
cryptography>=41.0.0
groq
